backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pandas==2.0.3
scikit-learn==1.3.0
supabase==2.3.0
python-dotenv==1.0.0
joblib==1.3.2
bcrypt==4.0.1
pyjwt==2.8.0
